Add LazaAdam and split embeddings.
