+++ merlin/models/tf/blocks/optimizer.py
@@ -16,6 +16,7 @@
 
 import collections
 import warnings
+from dataclasses import dataclass
 from typing import Callable, List, Optional, Sequence, Tuple, Union
 
 import numpy as np
@@ -28,6 +29,32 @@ from merlin.models.tf.utils import tf_utils
 
 Tensor = Union[tf.Tensor, tf.SparseTensor, tf.RaggedTensor]
 FloatTensorLike = Union[tf.Tensor, float, np.float16, np.float32, np.float64]
+
+
+@dataclass
+class OptimizerBlocks:
+    """dataclass for a pair of optimizer and blocks that the optimizer should apply to.
+    Example:
+        ml.OptimizerBlocks("sgd", [user_tower, third_tower])
+        ml.OptimizerBlocks("adam", item_tower)
+    """
+
+    optimizer: Union[str, tf.keras.optimizers.Optimizer]
+    blocks: Sequence[Block]
+
+    def get_config(self):
+        """return a tuple of serialized keras objects"""
+        return (
+            tf.keras.utils.serialize_keras_object(self.optimizer),
+            [tf.keras.utils.serialize_keras_object(block) for block in self.blocks],
+        )
+
+    @classmethod
+    def from_config(cls, config):
+        return cls(
+            tf.keras.optimizers.deserialize(config[0]),
+            [tf.keras.layers.deserialize(block) for block in config[1]],
+        )
 
 
 @tf.keras.utils.register_keras_serializable(package="merlin.models")
@@ -43,22 +70,26 @@ class MultiOptimizer(tf.keras.optimizers.Optimizer):
 
     For example:
     ```python
-      user_tower = ml.InputBlock(schema.select_by_tag(Tags.USER), ml.MLPBlock([512, 256]))
-      item_tower = ml.InputBlock(schema.select_by_tag(Tags.ITEM), ml.MLPBlock([512, 256]))
-      third_tower = ml.InputBlock(schema.select_by_tag(Tags.ITEM), ml.MLPBlock([64]))
-      two_tower = ml.ParallelBlock({"user": user_tower, "item": item_tower})
-      model = ml.Model(two_tower, ml.ItemRetrievalTask())
-      optimizer = MultiOptimizer(default_optimizer="adam",
+      import merlin.models.tf as ml
+      user_tower = ml.InputBlock(schema.select_by_tag(Tags.USER)).connect(ml.MLPBlock([512, 256]))
+      item_tower = ml.InputBlock(schema.select_by_tag(Tags.ITEM)).connect(ml.MLPBlock([512, 256]))
+      third_tower = ml.InputBlock(schema.select_by_tag(Tags.ITEM)).connect(ml.MLPBlock([64]))
+      three_tower = ml.ParallelBlock({"user": user_tower, "item": item_tower, "third": third_tower})
+      model = ml.Model(three_tower,  ml.BinaryClassificationTask("click"))
+
+      # The third_tower would be assigned the default_optimizer ("adagrad" in this example)
+      optimizer = ml.MultiOptimizer(default_optimizer="adagrad",
         optimizers_and_blocks=[
-          (tf.keras.optimizers.SGD(), [user_tower, third_tower]),
-          (tf.keras.optimizers.Adam(), item_tower),
+          ml.OptimizerBlocks(tf.keras.optimizers.SGD(), user_tower),
+          ml.OptimizerBlocks(tf.keras.optimizers.Adam(), item_tower),
         ])
 
-      # The string identification of optimizer is also acceptable:
-      optimizer = MultiOptimizer(default_optimizer="adam",
+      # The string identification of optimizer is also acceptable, here "sgd" for the third_tower
+      # The variables of BinaryClassificationTask("click") would still use the default_optimizer
+      optimizer = ml.MultiOptimizer(default_optimizer="adam",
         optimizers_and_blocks=[
-          ("sgd", [user_tower, third_tower]),
-          ("adam", item_tower),
+          ml.OptimizerBlocks("sgd", [user_tower, third_tower]),
+          ml.OptimizerBlocks("adam", item_tower),
         ])
     ```
 
@@ -66,21 +97,24 @@ class MultiOptimizer(tf.keras.optimizers.Optimizer):
 
     def __init__(
         self,
-        optimizers_and_blocks: Sequence[
-            Tuple[Union[str, tf.keras.optimizers.Optimizer], Sequence[Block]]
-        ],
+        optimizers_and_blocks: Sequence[OptimizerBlocks],
         default_optimizer: Union[str, tf.keras.optimizers.Optimizer] = "rmsprop",
         name: str = "MultiOptimizer",
+        **kwargs,
     ):
-        """Initializes an CompositeOptimizer instance.
+        """Initializes an MultiOptimizer instance.
 
         Parameters
         ----------
-        optimizers_and_blocks:  List of tuples of (optimizer instance, blocks that the optimizer
-            should apply to, which could be a list of blocks or only one block).
-        default: default optimizer for all blocks not specified in optimizers_and_blocks, by
+        optimizers_and_blocks: Sequence[OptimizerBlocks]
+            List of OptimizerBlocks(dataclass), the OptimizerBlocks contains two items, one is
+            optimizer, another one is a list of blocks or a block that the optimizer should apply
+            to. See 'class OptimizerBlocks'
+        default_optimizer: Union[str, tf.keras.optimizers.Optimizer]
+            Default optimizer for the rest variables not specified in optimizers_and_blocks, by
             default "rmsprop".
-        name: The optimizer name.
+        name:str
+            The name of MultiOptimizer.
         """
         super().__init__(name=name)
         self.name = name
@@ -88,16 +122,21 @@ class MultiOptimizer(tf.keras.optimizers.Optimizer):
             raise ValueError("`optimizers_and_blocks` can't be empty")
         self.default_optimizer = tf.keras.optimizers.get(default_optimizer)
         self.optimizers_and_blocks = []
-        for i, optimizer_and_blocks in enumerate(optimizers_and_blocks):
-            optimizer = tf.keras.optimizers.get(optimizer_and_blocks[0])
-            self._track_trackable(optimizer, name=f"Optimizer{i}")
-            self.optimizers_and_blocks.append((optimizer, optimizer_and_blocks[1]))
-        self.update_optimzier_and_block = []
+        for i, pair in enumerate(optimizers_and_blocks):
+            pair.optimizer = tf.keras.optimizers.get(pair.optimizer)
+            self._track_trackable(pair.optimizer, name=f"Optimizer{i}")
+            self.optimizers_and_blocks.append(pair)
+        if "update_optimzier_and_block" in kwargs:
+            # only for from_config, where there is already self.update_optimzier_and_block
+            self.update_optimzier_and_block = kwargs.get("update_optimzier_and_block")
+        else:
+            self.update_optimzier_and_block = []
 
     def _get_trainable_variables_optimizer_dict(self, optimizers_and_blocks, require_disjoint=True):
-
         attribute = "_trainable_weights"
-        for optimizer, blocks in optimizers_and_blocks:
+        for pair in optimizers_and_blocks:
+            optimizer = pair.optimizer
+            blocks = pair.blocks
             for block in blocks:
                 # Iterate all submodule (BFS) except ModelContext
                 # Note: block.trainable_variables is not used because modelcontext contain all
@@ -110,12 +149,12 @@ class MultiOptimizer(tf.keras.optimizers.Optimizer):
                     current_module = deque.popleft()
                     if hasattr(current_module, attribute):
                         for v in current_module._trainable_weights:
-                            if require_disjoint and v.ref() in self.var_optimizer_dict:
+                            if require_disjoint and (v.ref() in self.var_optimizer_dict):
                                 raise ValueError(
                                     f"The set of variables handled by each optimizer should be "
                                     f"disjoint, but variable {v} of {current_module} in block "
                                     f"{block} is handled both by {self.var_optimizer_dict[v.ref()]}"
-                                    f" and {optimizer}."
+                                    f"and {optimizer}."
                                 )
                             self.var_optimizer_dict[v.ref()] = optimizer
 
@@ -158,47 +197,47 @@ class MultiOptimizer(tf.keras.optimizers.Optimizer):
                 name=name,
                 experimental_aggregate_gradients=experimental_aggregate_gradients,
             )
-        return
 
     def add(
         self,
-        optimizer: Union[str, tf.keras.optimizers.Optimizer],
-        blocks: Sequence[Block],
+        optimizer_blocks: OptimizerBlocks,
     ):
-        """add another optimzier and specify which block to apply this optimizer to. If the block is
-        specified with an optimizer before, it may raise conflict error, please consider calling
-        self.update() method"""
+        """add another optimzier and specify which block to apply this optimizer to"""
         len_exist_optimizers = len(self.optimizers_and_blocks)
-        optimizer = tf.keras.optimizers.get(optimizer)
+        optimizer_blocks.optimizer = tf.keras.optimizers.get(optimizer_blocks.optimizer)
+        optimizer = optimizer_blocks.optimizer
         # Check if already track the optimizer
         optimizer_not_exists = True
-        for opt, blocks in self.optimizers_and_blocks:
-            if optimizer == opt:
+        for opt_blocks in self.optimizers_and_blocks:
+            if optimizer == opt_blocks.optimizer:
                 optimizer_not_exists = False
         if optimizer_not_exists:
             self._track_trackable(optimizer, name=f"Optimizer{1+len_exist_optimizers}")
 
-        self.optimizers_and_blocks.append((optimizer, blocks))
+        self.optimizers_and_blocks.append(optimizer_blocks)
         return
 
     def update(
         self,
-        optimizer: Union[str, tf.keras.optimizers.Optimizer],
-        blocks: Sequence[Block],
+        optimizer_blocks: OptimizerBlocks,
     ):
         """update the optimzier of a block, which would update the block's optimizer no matter
         what optimizer it used to utilize. If the block is not specified with an optimizer before,
-        this functions would have the same functionality as self.add()"""
+        this functions would have the same functionality as self.add()
+
+        Note: the optimizer_blocks would be kept in self.update_optimzier_and_blocks, instead of
+        self.optimizers_and_blocks"""
         len_exist_optimizers = len(self.optimizers_and_blocks)
+        optimizer, blocks = optimizer_blocks.optimizer, optimizer_blocks.blocks
         optimizer = tf.keras.optimizers.get(optimizer)
         # Check if already track the optimizer
         optimizer_not_exists = True
-        for opt, pre_blocks in enumerate(self.optimizers_and_blocks):
-            if optimizer == opt:
+        for pair in self.optimizers_and_blocks:
+            if optimizer == pair.optimizer:
                 optimizer_not_exists = False
-            for pre_block in pre_blocks:
+            for pre_block in pair.blocks:
                 if pre_block in blocks:
-                    pre_blocks.remove(pre_block)
+                    pair.blocks.remove(pre_block)
         if optimizer_not_exists:
             self._track_trackable(optimizer, name=f"Optimizer{1+len_exist_optimizers}")
         self.update_optimzier_and_blocks.append((optimizer, blocks))
@@ -209,40 +248,38 @@ class MultiOptimizer(tf.keras.optimizers.Optimizer):
         config = tf_utils.maybe_serialize_keras_objects(self, config, ["default_optimizer"])
         config["name"] = self.name
         config["optimizers_and_blocks"] = []
-        for optimizer, blocks in self.optimizers_and_blocks:
-            config["optimizers_and_blocks"].append(
-                (
-                    tf.keras.utils.serialize_keras_object(optimizer),
-                    [tf.keras.utils.serialize_keras_object(block) for block in blocks],
-                )
-            )
+        for optimizer_blocks in self.optimizers_and_blocks:
+            config["optimizers_and_blocks"].append(optimizer_blocks.get_config())
+        for optimizer_blocks in self.update_optimzier_and_block:
+            config["update_optimizers_and_blocks"].append(optimizer_blocks.get_config())
         return config
 
     @classmethod
     def from_config(cls, config):
         config["default_optimizer"] = tf.keras.optimizers.deserialize(config["default_optimizer"])
-        optimizers_and_blocks = []
-        for optimizer, blocks in config["optimizers_and_blocks"]:
-            optimizers_and_blocks.append(
-                (
-                    tf.keras.optimizers.deserialize(optimizer),
-                    [tf.keras.layers.deserialize(block) for block in blocks],
-                )
-            )
+        optimizers_and_blocks, update_optimizers_and_blocks = [], []
+        for optimizer_blocks_config in config["optimizers_and_blocks"]:
+            optimizers_and_blocks.append(OptimizerBlocks.from_config(optimizer_blocks_config))
         config.update({"optimizers_and_blocks": optimizers_and_blocks})
+        if config.get("update_optimizers_and_blocks"):
+            for optimizer_blocks_config in config["update_optimizers_and_blocks"]:
+                update_optimizers_and_blocks.append(
+                    OptimizerBlocks.from_config(optimizer_blocks_config)
+                )
+        config.update({"update_optimizers_and_blocks": update_optimizers_and_blocks})
         return cls(**config)
 
     @property
     def iterations(self):
         """See base class."""
         # Returning iterations from the first optimizer.
-        return self.optimizers_and_blocks[0][0].iterations
+        return self.optimizers_and_blocks[0].optimizer.iterations
 
     @iterations.setter
     def iterations(self, variable):
         """See base class."""
-        for optimizer, _ in self.optimizers_and_blocks:
-            optimizer.iterations = variable
+        for optimizer_blocks in self.optimizers_and_blocks:
+            optimizer_blocks.optimizer.iterations = variable
 
     def variables(self):
         """Returns the optimizer's variables."""
@@ -253,14 +290,15 @@ class MultiOptimizer(tf.keras.optimizers.Optimizer):
     def weights(self) -> List[tf.Variable]:
         """Returns the optimizer's variables."""
         weights = []
-        for optimizer, _ in self.optimizers_and_blocks:
-            weights += optimizer.weights
+        for optimizer_blocks in self.optimizers_and_blocks:
+            weights += optimizer_blocks.optimizer.weights
         return weights
 
     @property
     def optimizers(self) -> List[tf.keras.optimizers.Optimizer]:
-        """Returns the optimizers in composite optimizer (in the original order)."""
-        return [optimizer for optimizer, _ in self.optimizers_and_blocks]
+        """Returns the optimizers in MultiOptimizer (in the original order). Note: default_optimizer
+        is included here"""
+        return [pair.optimizer for pair in self.optimizers_and_blocks] + [self.default_optimizer]
 
 
 @tf.keras.utils.register_keras_serializable(package="merlin.models")
